Hotfix for YUM repos w/o Signature checking
As the Signature checks are currently failing https://github.com/sapcc/swift-http-import/issues/49,
the signature check was disbaled in the config. For YUM repos that excludes the signature file
repodata/repomd.xml.asc from being transfered.
To avoid issues with e.g. zypper the transfer of this signature is enabled with this commit again,
although it is not validated.