revert from github.com/ProtonMail/go-crypto to golang.org/x/crypto/openpgp
The ProtonMail fork has removed support for v3 signatures, which is a
reasonable choice for modern libraries since v4 and v5 signatures are
much more secure, but lots of SLES 12 repos have old v3 signatures, so
we need to continue supporting those for the rest of SLES 12's lifecycle.

In a related change, we now also cache expired keys. The signatures on
mny of these SLES 12 repos were made with expired keys, but since the
signature was made before the key expired, the signature is still fine.

/cc @talal