github-releases: do not list release assets
The `github.RepositoryRelease` struct already has a slice of its release
assets therefore we don't need to make a separate API request to get a
list of release assets.

Additionally, we construct the `FileSpec.Path` manually as the JSON
response from GitHub's API will escape characters like `+` in the tag
name therefore we can not rely on `BrowserDownloadURL` field and instead
have to construct the path using `TagName` field instead.